% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{chr2list}
\alias{chr2list}
\title{Parse a character string to a list}
\usage{
chr2list(chr)
}
\arguments{
\item{chr}{A character string that represents a list or a vector
see Notes for details.}
}
\value{
A list or a vector
}
\description{
These function parsers characters into list and
  list into characters. They help to parse the argument and makes the
  user input easier and intutive.
}
\note{
A semicolon in the string separates the elements of a list while
  A comma in the string separates the elements of a vector. See example.
}
\examples{
chr2list("1, 2, 3") # c(1, 2, 3)
chr2list("1, 2; 3, 4") # list(c(1, 2), c(3, 4))
chr2list("1:2; 3:7") # list(1:2, 3:7)
chr2list("1, 2, 3;") # list(c(1, 2, 3))
}
