% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{depth}
\alias{depth}
\title{Calculates the depth of a list}
\usage{
depth(this, thisdepth = 1)
}
\arguments{
\item{this}{A list whose depth needs to be calculated}

\item{thisdepth}{The default value if the list is flat}
}
\value{
An integer specifying the depth of a list
}
\description{
Calculates the depth of a list
}
\examples{
depth(list(list(1, 2), list(2, 3))) # 3
depth(1:5) # 1
depth(list(list(1, 2), list(2, list(3, 4)))) # 4
depth(list(1:2, 3:4)) # 2
}
\seealso{
https://stackoverflow.com/a/13433689/1987629 (Function copied from)
}
