% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{list2chr}
\alias{list2chr}
\title{Parse List/Vector to a character string}
\usage{
list2chr(list)
}
\arguments{
\item{list}{A list or a vector which need to be deparsed as a string}
}
\value{
A character string representing the list
}
\description{
Parse List/Vector to a character string
}
\note{
This is the reverse for of @seealso chr2list
}
\examples{
list2chr(c(1, 2, 3)) # "1; 2; 3"
list2chr(list(c(1, 2), c(3, 4))) # "1, 2; 3, 4"
list2chr(list(1:2, 3:7)) # "1, 2; 3, 4, 5, 6, 7"
}
